% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_logistic_derivative.R
\name{calculate_logistic_derivative}
\alias{calculate_logistic_derivative}
\title{Calculate the Rate of Change of a Logistic Population}
\usage{
calculate_logistic_derivative(t, P, parms)
}
\arguments{
\item{t}{Numeric (optionally a vector); the time steps to evaluate at. Note
that the logistic ODE is autonomous, but this is required for deSolve
compatibility.}

\item{P}{Numeric (optionally a vector); the population size(s) at which the
derivative should be evaluated.}

\item{parms}{Named list containing the entries "r" and "K", both positive
numeric values. The list MUST be of the form list("r" = #, "K" = #), but
the order does not matter. Formatting is specific for deSolve
compatibility, so it has to be like this until a tidyverse version of
deSolve is released, probably.}
}
\value{
A list containing the calculated derivatives.
}
\description{
Calculates the value of dP/dt when the population is at size P.
This assumes the population is growing logistically with know parameters r,
the intrinsic growth rate, and K, the carrying capacity.
}
\details{
Designed for use with the package deSolve.
}
\examples{
calculate_logistic_derivative(10, 50, list("r" = 0.1, "K" = 100))
}
