% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validate.R
\name{cross_validate}
\alias{cross_validate}
\title{k-Fold cross validation for OLS model with smoothing}
\usage{
cross_validate(data, k = 3, time_step = NULL, reduced = TRUE, lambda = 0)
}
\arguments{
\item{data}{A data frame of logistic growth time series data, one column
should be named "t" and one should be named "P".}

\item{k}{Numeric; the number of partitions for k-fold cross validation.
defaults to 3 in accordance with Stein et al's recommendation.}

\item{time_step}{Numeric; also strictly positive. Can be inputted manually
but defaults to the maximum time value divided by the number of time
points used.}

\item{reduced}{Logical; if TRUE, the carrying capacity is assumed to be 1 and
only r is fitted in the model. If FALSE, both K and r are fitted.}

\item{lambda}{Numeric; strictly positive. The smoothing parameter to use
for Tikhonov regularization in the regression model. Defaults to zero.}
}
\value{
The cross-validation error for k-fold cross validation on the model.
}
\description{
Performs cross validation using the method of Stein et al., "Ecological
Moldeing from Time Series Infernece: Insight into Dynamics and Stability of
Intestinal Microbiota; PLoS Comp Bio 2013.
}
\details{
Stein et al. recommended performing 3-fold cross validation 10 times. This
should be accomplished fairly easily with the `replicate` command.
}
