% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_logistic_data.R
\name{model_logistic_data}
\alias{model_logistic_data}
\title{Apply a Discretized OLS model to logistic time series data}
\usage{
model_logistic_data(
  df,
  smoothing = 0,
  dimensionless = TRUE,
  make_plot = FALSE,
  print_res = FALSE,
  ts = NULL
)
}
\arguments{
\item{df}{A data frame; typically generated by generate_data_frame(). One
column should be named "t" and numeric, and the second column should be
named "P" and numeric.}

\item{smoothing}{Nonnegative numeric; if >0 applies Tikhonov regularization
to the OLS model where this value is lambda. For our purposes, this
essentially forces the estimations to be smaller.}

\item{dimensionless}{Logical; if TRUE, the model is assumed to have one
parameter, r, which is estimated. If FALSE, the model is assumed to have
two parameters, r and K, which are estimated.}

\item{make_plot}{Logical; if TRUE, prints a plot of the linearized data.}

\item{print_res}{Logical; if TRUE, the values of the estimated parameters
are printed to the console.}

\item{ts}{Numeric; the step between values of time. If not specified, this is
calculated as the max time over the number of times.}
}
\value{
A named numeric vector of parameters which were estimated.
}
\description{
An Euler discretization of the logistic growth ODE problem is used to
formulate the problem as an Ordinary Least Squares regression problem. So,
OLS is used to estimate the value of the logistic growth parameters, r, and
optionally K if the model has not been normalized.
}
\details{
Applying Tikhonov regularization (AKA "Ridge regression") is also supported
by specifying the value of the smoothing constant. Options are provided to
print a plot of linearized data and print the results to the console.
}
