% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_logistic_equation.R
\name{solve_logistic_equation}
\alias{solve_logistic_equation}
\title{Analytic Solution to the Logistic Equation}
\usage{
solve_logistic_equation(P0, K, r, t)
}
\arguments{
\item{P0}{Real scalar; the initial population value.}

\item{K}{Real scalar; the carrying capacity.}

\item{r}{Real scalar; the intrinsic population growth rate.}

\item{t}{Real scalar; the time point at which to evaluate the population.}
}
\value{
Real scalar; the size of the population with the given parameters at
    time t.
}
\description{
Uses the known analytic solution to the one-population logistic growth
differential equation to calculate the size of a population at a specific
time point, given certain population parameters.
}
\examples{
solve_logistic_equation(100, 1000, 0.1, 10)
}
